requirements.txt
langchain==0.3.25
langchain-google-genai==2.1.4
langchain-groq==0.2.1
langchain-community==0.3.24
langchain-core==0.3.60
pypdf==5.1.0
python-dotenv==1.0.1
colorama==0.4.6
python-pptx==1.0.2
python-docx==1.1.2
streamlit==1.42.0